Move RegistrySpy
