+++ Tests/PlayMode/Mocks/RegistrySpy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 61ad113d8b55e994394ff1f323d10587
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
